terminando de acomodar cosas y ademas aregue una lista en el dto de artist para q tenga las canciones

+++ music-app/src/main/java/com/musicspring/app/music_app/artist/service/ArtistService.java
@@ -2,12 +2,16 @@ package com.musicspring.app.music_app.artist.service;
 
 import com.musicspring.app.music_app.artist.model.entities.ArtistEntity;
 import com.musicspring.app.music_app.artist.repository.ArtistRepository;
+import com.musicspring.app.music_app.shared.IService;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 
 import java.util.List;
 import java.util.Optional;
 
-public class ArtistService {
+public class ArtistService implements IService<ArtistEntity> {
     private final ArtistRepository artistRepository;
 
     @Autowired
@@ -15,20 +19,30 @@ public class ArtistService {
         this.artistRepository = artistRepository;
     }
 
-    public List<ArtistEntity> getAll(){
-        return artistRepository.findAll();
-    }
-
-    public Optional<ArtistEntity> getbyId(Long id){
-        return artistRepository.findById(id);
-    }
-
-    public ArtistEntity save(ArtistEntity artist){
-        return artistRepository.save(artist);
-
-    }
-
-    public void delete(Long id){
+    @Override
+    public void deleteById(Long id) {
+        if (!artistRepository.existsById(id)) {
+            throw new EntityNotFoundException("Artist with id " + id + " not found");
+        }
         artistRepository.deleteById(id);
+    }
+
+    @Override
+    public Page<ArtistEntity> findAll(Pageable pageable) {
+        return artistRepository.findAll(pageable);
+    }
+
+    @Override
+    public ArtistEntity findById(Long id) {
+        return artistRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Artist with id " + id + " not found"));
+    }
+
+    @Override
+    public ArtistEntity save(ArtistEntity artistEntity) {
+        if(artistEntity == null) {
+            throw new IllegalArgumentException("ArtistEntity cannot be null");
+        }
+        return artistRepository.save(artistEntity);
     }
 }
